1.16

+++ smart_pantry/lib/features/add_item/screens/add_item_form_screen.dart
@@ -91,7 +91,7 @@ class _AddItemFormScreenState extends State<AddItemFormScreen> {
         emoji: _selectedCategory.emoji,
         purchaseDate: DateTime.now(),
         expiryDate: _expiryDate,
-        quantity: _quantity,
+        quantity: _quantity.toDouble(),
         unit: _selectedUnit,
       );
